updated backend testing

+++ client/src/components/ProjectCard.jsx
@@ -3,11 +3,10 @@ import { Typography, Card, Accordion, AccordionSummary, AccordionDetails, CardMe
 import { useState } from "react";
 import { GetIcon, GetTechIcon, GetIconButton } from "./Icon";
 import { useSelector, useDispatch } from 'react-redux'
-import { palette } from "../styles/theme";
 import { deleteProject } from "../store/projectSlice";
 import { useNotification } from '../contexts/NotificationContext'
 
-const cardStyle = {
+const sxProjectCard = {
   card: {
     display: 'flex',
     flexDirection: 'column',
@@ -95,6 +94,7 @@ const ProjectCard = ({ project, index }) => {
 
     const handleDelete = async () => {
       if (!user) return
+      if (!window.confirm(`Are you sure you want to delete "${project.title}"?`)) return
       try {
         await dispatch(deleteProject(project.id)).unwrap()
         showNotification('info', `${project.title.charAt(0).toUpperCase() + project.title.slice(1)} deleted`)
@@ -104,9 +104,9 @@ const ProjectCard = ({ project, index }) => {
     }
 
     return (
-        <Card sx={cardStyle.card}>
+        <Card sx={sxProjectCard.card}>
 
-            <Box sx={cardStyle.projectNo}>
+            <Box sx={sxProjectCard.projectNo}>
                 {user && <GetIconButton title={`delete ${project.title}`} type='delete' onClick={handleDelete} />}
                 <GetIconButton title='link: demo site' type='demolink' href={project.demoUrl} target='_blank' />
                 <GetIconButton title='link: github repository' type='github' href={project.githubUrl} target='_blank' />
@@ -117,7 +117,7 @@ const ProjectCard = ({ project, index }) => {
                 component="img"
                 image={project.imagePath}
                 alt={project.title}
-                sx={cardStyle.projectImage}
+                sx={sxProjectCard.projectImage}
             />
 
             <Accordion
@@ -136,10 +136,10 @@ const ProjectCard = ({ project, index }) => {
                     sx={{ px: 0 }}
                     >
 
-                    <Box sx={cardStyle.projectTitleRow} width="100%">
-                        <Box sx={cardStyle.techList}>
+                    <Box sx={sxProjectCard.projectTitleRow} width="100%">
+                        <Box sx={sxProjectCard.techList}>
                             {project.tech.map((tech) => (
-                              <GetTechIcon key={tech.id} className={tech.icon} size={25} sx={cardStyle.techIcon} title={tech.name} />
+                              <GetTechIcon key={tech.id} className={tech.icon} size={25} sx={sxProjectCard.techIcon} title={tech.name} />
                             ))}
                         </Box>
 
@@ -158,7 +158,7 @@ const ProjectCard = ({ project, index }) => {
                 </AccordionSummary>
 
                 <AccordionDetails sx={{ px: 0 }}>
-                    <Box sx={cardStyle.projectDesc}>
+                    <Box sx={sxProjectCard.projectDesc}>
                         <Typography variant="body2">{project.description}</Typography>
                     </Box>
                 </AccordionDetails>
